+++ activestorage/CHANGELOG.md
@@ -5,6 +5,11 @@
     *Don Sisco*
 
 
+## Rails 6.1.4.4 (December 15, 2021) ##
+
+*   No changes.
+
+
 ## Rails 6.1.4.3 (December 14, 2021) ##
 
 *   No changes.